Make a loop not copy
clang-tidy warned that we're taking a string by copy.  I'm not sure why
it wasn't able to find this earlier, but alas, this is to make the build
pass again.

Tested:
I don't have a system that implements this, so I cant functionally test,
but this is a trivial change.  Clang-tidy now passes.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I333ae27a0d331c461c0aefc4f2c3dda60749e073