Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 2:
> > 
> > Hi JunLin, Can this also persist the event subscriber's id ? similar to how we persist the session-id
> 
> I try to add the code for persist id. But still maintains the behavior that the subscription id will be non-sequential when it is created.
> Does need to change the id based on the creation such like subscription/1, subscription/2, subscription/3?

I was referring to the id which is for example 777063716 in the subscriber url /redfish/v1/EventService/Subscriptions/777063716
This id needs to be persisted, and should not change after a BMC reboot or bmcweb restart. This need not be a sequential number.

Patch-set: 3
Reviewer: Gerrit User 1000382 <1000382@adbd0a64-1f21-4d83-b585-671fe73cb6e4>