Add support for the BootSourceOverrideMode
BootSourceOverrideMode is the BIOS boot mode to use when
the system boots from the BootSourceOverrideTarget boot source.

Possible values are:
-Legacy - The system boots in non-UEFI boot mode to
          the boot source override target
-UEFI   - The system boots in UEFI boot mode to
          the boot source override target

For the architectures that don't implement Legacy/UEFI boot mode
property (for example POWER), don't publish BootSourceOverrideMode
interface at all.

Redfish validator has been executed on this change and no new
error was found.

Also tested with the help of the openbmc-test-automation framework.
When the Legacy/UEFI boot mode is present in the system command for
testing is:
  robot -v PLATFORM_ARCH_TYPE:x86 \
        -v OPENBMC_HOST:<BMC IP> \
        redfish/systems/test_boot_devices.robot

When the Legacy/UEFI boot mode is not present in the system command
for testing is:
  robot -v OPENBMC_HOST:<BMC IP> \
        redfish/systems/test_boot_devices.robot

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>
Change-Id: Id64e428752b820d481ce15d56566dd7b06738ec0