Add support for the BootSourceOverrideMode
BootSourceOverrideMode is the BIOS boot mode to use when
the system boots from the BootSourceOverrideTarget boot source.

Possible values are:
-Legacy - The system boots in non-UEFI boot mode to
          the boot source override target
-UEFI   - The system boots in UEFI boot mode to
          the boot source override target

Redfish validator has been executed on this change and no new
error was found.

This change is dependant on the "Add BootType to the default settings"
patch (https://gerrit.openbmc-project.xyz/c/openbmc/openbmc/+/40910)
to the 'meta-phosphor' project.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>
Change-Id: Id64e428752b820d481ce15d56566dd7b06738ec0