Update patch set 5
Patch Set 5:

This was discussed on discord as for why this is accepted but https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/38044 isn't. This line is 

"Is this expected to possibly be empty on a functional and working system"

i.e. 38044 was trying to remove empty properties that were expected on working system

In the future, there might be implementations that don't have all those inventory properties set (e.g. SparePartNumber) so SparePartNumber would fall on the other side of the line and would not be displayed if empty

Patch-set: 5
Reviewer: Gerrit User 1000020 <1000020@adbd0a64-1f21-4d83-b585-671fe73cb6e4>