Expose Processor PPIN property
Some Processors may support exposing a globally unique identifier in
addition to a serial number. Redfish supports this via
ProcessorId.ProtectedIdentificationNumber property. ("Protected"
indicates it may be a sensitive value due to tracking/privacy concerns -
Intel CPUs allow the customer to mask the PPIN so it can't be exposed.)

cpuinfoapp in smbios-mdr repo is currently implementing this interface.

Tested:
- New property is shown (only for cpus with valid PPIN)
- Passed Redfish Service Validator with no errors.

Signed-off-by: Jonathan Doman <jonathan.doman@intel.com>
Change-Id: I233931ac640338e8f50f37ba4cc7298c7a0db9bf