Make missing static files directory non-fatal
Today, bmcweb requires the /usr/share/www directory to exist.  In most
cases where bmcweb was installed with make install, this is fine, but in
development scenarios, we'd like to be able to boot.

This commit moves what used to be a fatal error (the directory not
existing) to a warning, to help with a common development issue I've
seen before.

Tested:
Launched without a /usr/share/www directory present, and bmcweb launched
successfully and did not throw an exception.

Change-Id: I775fafd86a4e2eef0bf73836d31a78fb320b61c0
Signed-off-by: Ed Tanous <edtanous@google.com>