Implements ThermalMetrics schema
The ThermalMetrics schema is a resource in Redfish version 2022.2[1].
It contains an array of temperature readings.
It is a child of ThermalSubsystem schema[2] and it represents the
thermal metrics of a chassis.
Reading the current value of each temperature sensor and the
corresponding link enumeration will be implemented in the next patch.

This commit implements the Get and Head methods of the Redfish
ThermalMetrics schema and implemented the basic information of Get.

[1] https://www.dmtf.org/sites/default/files/standards/documents/DSP0268_2022.2.pdf
[2] https://redfish.dmtf.org/schemas/v1/ThermalMetrics.v1_0_1.json

Test:
1. Validator passed.
2. doGet method:
"""
curl -k -H "X-Auth-Token: $token" -X GET https://${bmc}/redfish/v1/Chassis/chassis/ThermalSubsystem/ThermalMetrics
{
  "@odata.id": "/redfish/v1/Chassis/chassis/ThermalSubsystem/ThermalMetrics",
  "@odata.type": "#ThermalMetrics.v1_0_1.ThermalMetrics",
  "Id": "ThermalMetrics",
  "Name": "Thermal Metrics",
}
"""
3. A bad chassis ID:
"""
curl -k -H "X-Auth-Token: $token" -X GET https://${bmc}/redfish/v1/Chassis/chassisBAD/ThermalSubsystem/ThermalMetrics
{
  "error": {
    "@Message.ExtendedInfo": [
    {
      "@odata.type": "#Message.v1_1_1.Message",
      "Message": "The requested resource of type Chassis named 'chassisBAD' was not found.",
      "MessageArgs": [
        "Chassis",
        "chassisBAD"
      ],
      "MessageId": "Base.1.13.0.ResourceNotFound",
      "MessageSeverity": "Critical",
      "Resolution": "Provide a valid resource identifier and resubmit the request."
    }
  ],
  "code": "Base.1.13.0.ResourceNotFound",
  "message": "The requested resource of type Chassis named 'chassisBAD' was not found."
  }
}
"""

Signed-off-by: zhanghaicheng <zhanghch05@inspur.com>
Change-Id: Ib4182e7dc6e204371636a33a391e8e2a58dad113