Remove default.css from the image
This is primarily to save space, and remove cruft.  As part of the "old"
json->html parser, this file was used to do the syntax highlighting for
the hilight-js class we used.  That parser has been removed, so this
code is now dead.

Tested:
Has no external impact to test.  Code builds.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Iebbf97003d41818abc2a9eae68c23c636f13d2e6