Update patch set 23
Patch Set 23:

Chicago : From this, how can we get a corresponding backend D-Bus object for a particular resource in a slot ?.. What is the key for GET and PATCH ?

Also, in Assembly, we had initially used "member_id" to represent each instance.. However, looks like it's a violation of the spec.  Assembly is taking some rework. Please help see if pcieslot also has to do something like that ?.

https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/39574

Patch-set: 23