Update Crashdump to the latest LogService and LogEntry schemas
This updates Crashdump to the latest LogService schema which now
has a "CollectDiagnosticData" Action that replaces the OEM actions.

It also updates to the latest LogEntry to use the new
"AdditionalDataURI" property to provide the Crashdump file URI.

Tested:
Confirmed that the Crashdump LogEntries are displayed and can
be collected correctly.

Confirmed that the schema changes pass the Redfish Service Validator.

Change-Id: I603ee40128a112f0910db3adf17f7510be45a1d6
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>