Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What is wrong with this as is? 
> I would think we would want individual options for things we are deprecating.

I don't understand the question. 
Is "this" the code prior to this commit request?
Is "this" the commit I've submitted for review?
Is "this" the addition of the "disable/enable the PATCH action"?
Is this question to me, to Ed, or all reviewers?

Patch-set: 2
Reviewer: Gerrit User 1000278 <1000278@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=0