Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Sent something to the mailing list here: https://lists.ozlabs.org/pipermail/openbmc/2020-December/024296.html
> 
> The reply on the mailing list made this seem, this was added on accident and the author didn't know of any clients that used this.

I see that, and that significantly changes the disposition of this patch, considering we're reasonably sure there are no active users, and this patch has sat on all the communication mediums for long enough that clients that care would've seen it and had a chance to comment on it.

Lets get a comment added the option saying we're going to remove it in ~4 months (my number is relatively arbitrary and subject to discussion if anyone wants to change it), then I suspect this can be good to merge.

Thanks Jonathan.

Patch-set: 2
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=0