Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What is wrong with this as is? 
> I would think we would want individual options for things we are deprecating.

I'm absent on checking in on this, but it was for some good reasons.

Initially, I didn't want to maintain a never ending swath of options, because we're already failing keeping things compiling with the options we have, and doubling the number of compile-time options seems like it's just asking for worse problems.

With that said, in the meantime since this patch was first pushed, I've had a chance to go look into that and come up with a better plan for how to have lots of compile time options while keeping things relatively sane and build able (see the XSS protection commit that now uses consexpr).  I think that gives us a path forward to be maintaining a lot more options in the future, and even though that work is ~5% done, I can at least see a path that's maintainable.

I still have some concerns over the core thought behind this which is the argument that we will follow the spec to the letter.  We definitely break the spec in places, or add additional features that are outside the resource spec in places where it's convenient.  I have a suspicion that nobody actually wants strict spec compliance, given the things that we'd have to remove or break to accomplish it (cookies, sensor override, ect).  I'm assuming most would prefer something like "Correctly written clients will see a correct implementation of Redfish, but we may extend it further than the spec allows" which at least thusfar, seems to be an ok happy medium between building things that work and still matching the intent of the spec.

With that all said, this patchset isn't important enough to hash that out for.  I'm fine with it in its current form.  Lets get documented when we plan to completely remove it, and this should be good to go.

Patch-set: 2
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=0