Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I have added the requested comment for the compile flag. I have chosen Q4 2021 as the timeframe for removal of the compile switch.

Great.

> 
> Ed, as the person who was responsible for making IE 9 work with the web server in Intel BMC generations I'm a little surprised you aren't more enthusiastic about having a uniform, well defined API. Every HTML4 server had some unique behavior requiring each browser to have code specifically crafted for the vagaries.

HTML is a good example of something that is still a pain today, because someone (microsoft) built an implementation, then people relied on that implementation.  If Microsoft had removed APIs from IE without warning like this patchset will do eventually, lots of things would've broken, and corporate websites would be way more broken than they already are.  That seems worse to me than the way history played out.

> 
> It's also odd to me that there are features/operations allowed that aren't part of the RF spec. Those kinds of items, to my perhaps naive thought process, should be exposed in the location RF set aside for them, the OEM blocks. The problem is that OpenBMC guidance is to eschew OEM blocks.

The problem here is that some of the things people want to do tend to go to the protocol level, not just the resource level, or have far reaching consequences.  If the OEM blocks were maintained and had sufficient quality, we'd have very different rules, but unfortunately, every one of them breaks the Redfish spec or guidance in subtle ways, hence the policy.  If you're stepping up to fix an maintain those, by all means, but the current policy is mostly there to stop the bleeding for the moment until we can regroup and figure something out that works in the long run.

If you think a different OEM policy is warranted, this is probably not the place to propose it;  The mailing list or the OEM review might be a good place to start.

Patch-set: 3
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>