Update patch set 3
Patch Set 3:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> Looks fine to me, although we should get those schemas deleted and message that this is a change that WILL break people.  With that said, I suspect they're all (or mostly) intel users, so I'll leave that rollout determination up to you.
> 
Downstream I'm going to maintain a patch with the current interface supported as deprecated for a while before I delete it.  I don't know if anyone else is using this interface right now.  I can ask on the mailing list before merging if that is the right approach.

> Ping me when the backend changes for this merge.
The backend is ready.  The D-Bus interfaces are the same, so I shouldn't need to update anything but my Redfish script.

Patch-set: 3