Update patch set 2
Patch Set 2:

On the system we tested this by running locally. The cookie is being blocked and we are no longer able to run locally. We can work around this by copying code on the server. I think we need to be able to set a flag for build time that disables this so we can run locally. Can we use the existing flags to determine if the set-cookie header is set or not?

Patch-set: 2
Reviewer: Gerrit User 1000281 <1000281@adbd0a64-1f21-4d83-b585-671fe73cb6e4>