Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > > Patch Set 2:
> > > 
> > > Lets get the option added then this LGTM.
> > 
> > Agree
> 
> Why? This doesn't break running locally.. I don't see a use case for why we need this excluded in insecure-disable-xss ? I.e. what is the use case this breaks?

Because at the time it was stated that it broke local development.  Considering this apparently doesn't (which is a little surprising, but I'll chalk that up to http magic I don't understand), it's fine as-is.

Patch-set: 3
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=+2