Add google service root
This commit introduces the following
=> Service root for Google
=> compiler option for the Google Root of Trust specific functionalities


Why is this change needed?
This change is going to help Google transition to use
Redfish to communicate with its root of trust devices.

There should be no physical impact on users outside of Google.

This functionality is guarded behind the compiler option:
-Dgoogle-api

this guard is needed to avoid adding unnecessary code that is no-op to
users outside of Google.

See https://gerrit.openbmc-project.xyz/c/openbmc/docs/+/41816 for
design doc on this feature.

Tested:
curl -vvvv --insecure --user $user_pass https://${bmc}/google/v1

Change-Id: I941b5cab55179279d0eff18aa29df62c3f226e47
Signed-off-by: Feras Aldahlawi <faldahlawi@gmail.com>