Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> @Gunnar  @Ed Tanous
> Is it feasible to use this method? Change the "doGet code" of the collection‘s member to a separate function, which is called separately after judging the "param".

Based on what I know, this wont work, and would require separation of the AsyncResp from the handler anyway.  If we're going to do that, we might as well just make the core RAII compliant as well, as it means that all the other filtering and expand parameters become easier in the future.

> Will too much code be modified? Or cause more problems?

Patch-set: 8
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>