Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > (1 comment)
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > (1 comment)
> > > > > 
> > > > > Looks fine to me, although we should get those schemas deleted and message that this is a change that WILL break people.  With that said, I suspect they're all (or mostly) intel users, so I'll leave that rollout determination up to you.
> > > 
> > > I'll leave that up to you.  The interface isn't the most self-describing, so I wouldn't assume anyone could've reverse engineered it without talking to you first.  Sending to the mailing list would be good just for visibility if someone comes out of the woodwork though.
> > > 
> > > > > 
> > > > Downstream I'm going to maintain a patch with the current interface supported as deprecated for a while before I delete it.  I don't know if anyone else is using this interface right now.  I can ask on the mailing list before merging if that is the right approach.
> > > 
> > > Cool.
> > > 
> > > > 
> > > > > Ping me when the backend changes for this merge.
> > > > The backend is ready.  The D-Bus interfaces are the same, so I shouldn't need to update anything but my Redfish script.
> > > 
> > > Can you point to the backend for this?  I thought PECI got removed from the openbmc kernel?
> > 
> > Oh, you mean public backend? 😊  The backend for Crashdump is all in Intel-BMC including PECI now that it was removed.  I can pull all of this code out into a downstream patch, if that is better.
> 
> That's kinda my point.... there's no way this could ever work as-written, but I'm not sure I want you to have to maintain a downstream patch like that either.  It'd be nice if you guys got the PECI patchset into mainline, but I'm assuming that ship has sailed and you guys decided it wasn't worth it, so I'm kind of at an impasse as to what to do in regards to this handler.

I think this may be why it's behind a build flag.

Last I heard internally, we were preparing for another run at PECI, but I don't know the status.  However, even if we get PECI into mainline, I don't know that we can get crashdump into mainline.  Crashdump is the host app for these D-Bus interfaces, so that middle layer will still be missing from upstream. 😞

It's my goal at some point to figure out if it's possible to either make my application fit in the Dump interface that was recently added or make the Dump interface more generic so it can fit my application as well.  Until then, I can carry this implementation as a downstream patch if we need to remove this build path.  I guess that may be good to ask on the mailing list as well to make sure nobody else is using this build option...

Patch-set: 3