Update patch set 57
Patch Set 57:

> I think you've misunderstood: two systems should not publish different message registries with the same name, namespace, and version.

Does that mean if the "OwningEntity" is changed to the company which owns this message registry then it would be fine?
We need to figure it out how it needs to be done.Random thought:  For time being can we have configure variable which is set to who owns the message regsitry..
Company can change this configure variable before building the image,


Name : I feel can be hardcoded here with the owning entity eg("Name": "CompanyName BIOS Attribute Registry")

RegistryVersion: If the owning entity is changed to the company who owns it then is there any problem if it is hardcoded.
If yes, Should it come from Dbus? 

Looking for your suggestion?

Patch-set: 57
Reviewer: Gerrit User 1000022 <1000022@adbd0a64-1f21-4d83-b585-671fe73cb6e4>