Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > > Patch Set 11:
> > > 
> > > > Patch Set 11:
> > > > 
> > > > > Patch Set 11:
> > > > > 
> > > > > Ed/Ratan
> > > > > I will work on this commit from next week onwards.
> > > > 
> > > > One thought I had, is I wonder if we overcomplicated the API by following the websockets model.   These are still transaction based, and only 1 wayso one brainstorming idea was to try to use the Response::write() method instead.  That would allow http keepalive to continue to work, wouldn't require us to invent a new timer system, and could mostly use the connection class as-is.
> > > > 
> > > > 
> > > > https://gerrit.openbmc-project.xyz/c/openbmc/bmcweb/+/39409
> > > > 
> > > > 
> > > > Food for thought, keep in mind, the above patchset is nowhere near working, or even compiling, but kinda goes through how I think it could work.
> > > 
> > > sure. will look into your commit and see if i can offload work
> > 
> > Ed
> > 
> > as discussed on the commit you mentioned, we cant use buffer body.
> > Can we proceed with dynamic_buffer_body and framework this commit implemented?
> 
> How do you plan to support keepalive and timeouts?  One thing this patch made me realize was that those two are kinda hard to implement in a model like this.  websockets largely avoids them because they have internal "ping" mechanisms, and are expected to stay alive forever, and therefore can have hard connection count limits.
> 
> > 
> > once we decide on buffer body type and framework, i will work incorporating outstanding comments on this commit.
> 
> sure.  I'm still of the opinion that buffer_body is the way to go here, but I'm interested to see if you can build something better with dynamic_body.



By going through with this commit, Can we divide the comments in following categories?

1) Flow Control

2) HTTP Response type(dynamic_body v/s buffer_body v/s string_body)

3) when to upgrade the connection?  Not needed as we should make it rule based

1) flow control
================

a) Read the data from the unix socket
     any error : Close both the ends.
b) write it on the http socket
     any error: close both the ends.

c) once successfully written then perform step a)

Are we good here?

2) HTTP Response type
=====================

buffer_body: It adds the chunked transfer encoding.
     cons: deprecated in http/2    
     
string_body or dynamic_body: I don't have any strong preference here.

In the existing bmcweb code we have the response class which is using the response type as 
string_body but can't be use as it is as it writes the content when code calls the
res.end().

in res.end() it calls the preaparePayload function which automatically adds the
content-length header with the size of the intermediate data and it is not correct
in this use case.

content-length should be the length of the full response, in this case (whole dump size).

We have two choices here.
1) Enhance the existing response class to support intermediate writes and don't depend on the
res.end() for writing.
2) Write the new class which has been done in the commit.

My preference to go with 1). Ed, are you with me or do you have other suggestion?


3) when to upgrade the connection?  Not needed as per my understanding if we make it rule based
==================================

We don't need this if we enhance the existing response and the connection class then it would be rule based.which function need to be called on connection and the response object , depends on the rule.

This is as per my understanding of this code, Ravi please correct me if I misunderstood.

Patch-set: 12
Reviewer: Gerrit User 1000022 <1000022@adbd0a64-1f21-4d83-b585-671fe73cb6e4>