Update patch set 57
Patch Set 57:

> Patch Set 57:
> 
> > I think you've misunderstood: two systems should not publish different message registries with the same name, namespace, and version.
> 
> Does that mean if the "OwningEntity" is changed to the company which owns this message registry then it would be fine?

It's not quite that easy, but the above would be a good start.  Defining how:
1. Versioning works
2. Namespacing between systems/CPUs/BIOSes work.

Is probably what's needed here.

> We need to figure it out how it needs to be done.Random thought:  For time being can we have configure variable which is set to who owns the message regsitry..

No.  That doesn't fully solve the issue.

> Company can change this configure variable before building the image,

That would imply that the registry parameters are driven by the BMC configuration, which they're not, they're driven by the BIOS that's installed.  Some mechanism to communicate that between the BMC and BIOS will be needed

> 
> 
> Name : I feel can be hardcoded here with the owning entity eg("Name": "CompanyName BIOS Attribute Registry")

I feel like it can't, as we shouldn't be hardcoding data like that in bmcweb.  If it needs to be dynamic, it needs an appropriate dbus interface.

> 
> RegistryVersion: If the owning entity is changed to the company who owns it then is there any problem if it is hardcoded.

Versioning and it changing dynamically changes are still a problem.

> If yes, Should it come from Dbus? 

I think so, but I haven't looked at the problem in depth.

> 
> Looking for your suggestion?

Patch-set: 57
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>