Add PerformService privilege
This commit creates a new OemOpenBMCPerformService privilege.

Background: OpenBMC implementations support sensitive operations which
may not be appropriate for a BMC administrator to use.  For example:
 - A manufacturing team may need to set a machine serial number or run
   host operations intended only for the manufacturing process.
 - A field service agent may need to set a MAC address after replacing
   a network FRU device, or to access the sensitive innards of the host.

The Redfish standard focuses on the BMC admin and operator experience
and does not support the use cases as identified above.  Redfish
refusals to add such functions to the standard are documented here:
 - https://redfishforum.com/thread/333/proposed-new-servicerep-role-privilege
 - https://redfishforum.com/thread/393/modify-schema-propertys-permission
 - https://redfishforum.com/thread/397/redfish-direction-update-eeproms

The Redfish solution is to allow Redfish services (such as BMCWeb) to
define OEM Privileges and custom Roles per the Redfish privilege model
described in the DSP0266 spec Security details > Authorization.  In
particular, subsection Privilege model > Roles.

The main idea is to protect access to sensitive operations by
requiring the OemOpenBMCPerformService privilege.  This privilege
should not be applied to standard Redfish operations; the Redfish-
supplied operation-to-privilege mapping should be used.  This
privilege is intended only for Redfish OEM operations and for
vendor-specific operations hosted on BMCWeb but outside of the
Redfish model.

Use cases:
1. In the default configuration, the Administrator role has all
   privileges, including the OemOpenBMCPerformService.  This ensures
   admin users are allowed to use all BMC operations.
2. To isolate service operations away from BMC admin users, the
   OemOpenBMCPerformService privilege can be removed from the
   Administrator role.  A custom ServiceAgent role would include the
   OemOpenBMCPerformService privilege.

This commit takes the first step toward enabling the second use case.

Tested:
- The admin user can continue to use the ConfigureUsers privilege.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>
Change-Id: Idaafdc42877c78fc6644687d14d567490d3f6fea