Add PerformService privilege
This commit creates a new OemOpenBMCPerformService privilege.

Background: OpenBMC implementations support sensitive operations which
may not be appropriate for a BMC administrator to use.  For example:
 - A manufacturing team may need to set a machine serial number or run
   host operations intended only for the manufacturing process.
 - A field service agent may need to set a MAC address after replacing
   a network FRU device, or to access the sensitive innards of the host.

The Redfish standard focuses on the BMC admin and operator experience
and does not support the use cases as identified above.  Redfish
refusals to add such functions to the standard are documented here:
 - https://redfishforum.com/thread/333/proposed-new-servicerep-role-privilege
 - https://redfishforum.com/thread/393/modify-schema-propertys-permission
 - https://redfishforum.com/thread/397/redfish-direction-update-eeproms

The Redfish solution is to allow Redfish services (such as BMCWeb) to
define OEM Privileges and custom Roles per the Redfish privilege model
described in the DSP0266 spec Security details > Authorization.  In
particular, subsection Privilege model > Roles.

The main idea is to protect access to sensitive operations by
requiring the OemOpenBMCPerformService privilege.  This privilege
should not be applied to standard Redfish operations.  The new
privilege is intended only for Redfish OEM operations and for
vendor-specific operations hosted on BMCWeb but outside of the
Redfish model.

To clarify, this commit only defines the new privilege.  It does not
add it to any operation-to-privilege maps.  And it only adds it to the
Administrator role.

Use cases:
1. The BMC Administrator has all privileges, including the
   OemOpenBMCPerformService, and they are allowed to use all BMC
   operations.  This is the default configuration.
2. The BMC Administrator has all privileges except for the service
   privileges, and the intention is admin users are not allowed to
   perform service operations.  A separate service account has the
   service privilege, for example, in a custom ServiceAgent role.

This commit takes the first step toward enabling the second use case.

Tested:
The admin user can continue to use the ConfigureUsers privilege.  for
example they can create and modify users.  Operators cannot.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>
Change-Id: Idaafdc42877c78fc6644687d14d567490d3f6fea