Allow AccountService members to be used to POST
DO NOT MERGE

The redfish specification states:
Submitting a POST request to a resource collection is equivalent to
submitting the same request to the Members property of that resource
collection.

So far as I'm aware, AccountService is the only collection today where
we support POST.  This commit makes us compliant, and allows parsing new
user data from either the POST or the first Member in the array.

At this point, if more than one Member is submitted, the code returns
bad request.  To support creation of multiple members would require
significant restructuring, which should be done, but not in the scope of
this commit.

Tested:
WIP

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Icf475d1891e9027a09b0f7d99abc5fbbd1a29b65