Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > I'm assuming this is WIP?
> 
> Yes, I'll mark it WIP.  I re-read the emails from Dec 2020 but have not yet looked at the statically-configurable (configurable at build-time) privilege registry (contrasted to the hard-coded privileges BMCWeb current has).
> 
> I introduced this OEM privilege to help me cleanly separate the OEM privileges (which we discussed in the email) from custom roles.  For example, a hypothetical custom Redfish ServiceAgent role would correspond to a custom priv-service phosphor privilege role, and also would include the this new OEM privilege.  I am writing a design doc for that, and it was just easier if I had a concrete OEM privilege to work with.

Unfortunately, because OEM privileges are..... OEM, by definition they will need to be configurable for everyone needs, so I'm not particularly inclined to approve a "concrete" OEM privilege.  Considering the current Redfish implementation implements the base privilege registry that's published by the standard, that should be available to consumers as an option (and probably the default option).

> 
> Changing topics, I believe the IBM specific rest APIs will follow the Redfish privilege/authorization model,

It'll be interesting to see how this works in practice.  I would prefer to not spread IBMisms through the generic privilege code, so please plan on making a generic API for this, that the IBM system can inject into to do its system-specific things.  In this way, we keep the company specific URLs contained, which should aboid having to write

if (ibm)
else if (google)
else if (intel)
else if (quanta)

in a bunch of places.

> using custom IBM roles and OemIBM privileges per the Redfish spec.  No new privilege system is needed.

I don't think the privilege system itself needs to change, it's perfectly capable of being dynamic, and is built at runtime based on the routes already.




PS, in general, I only put WIP on the title and topic, as marking things WIP in gerrit prevents any sort of discussion until the patchset is opened.

Patch-set: 3
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Work-in-progress: false