Update patch set 34
Patch Set 34:

> Patch Set 32:
> 
> > Patch Set 32:
> > 
> > @Gunnar @Ed
> > 
> > Performance testing (average of 5 times):
> > 
> > > curl -k -H "X-Auth-Token: $token" https://${bmc}/redfish/v1/Chassis/chassis/Sensors
> > 
> > Before: 0.275 s
> > With this change: 0.394 s

So a 43% performance drop for an API that is already slower than it should be.  That seems like a problem if this isn't a complete replacement for the thermal schemas.

> > 
> Cool. Thanks. Can you add this one to the commit msg? 
> Would vary depending on the number of sensors on your system.
> 
> > > curl -k -H "X-Auth-Token: $token" https://${bmc}/redfish/v1/Chassis/chassis/Sensors/ps0_input_current
> > 
> > Before: 0.874 s
> > With this change: 0.913 s
> 
> Not sure how a single sensor is useful information here. Was thinking you would:
> 1. start the time
> 2. call /redfish/v1/Chassis/chassis/Sensors
> 3. call each of the members in the sensor collection from the call before

Also call all the elements in the thermal schema.  When I've done this in the past, I've just run the service validator, it include request time.

> 4. end the time

Patch-set: 34
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>