Redfish: Fix the health and state properties of processor
The processor resource needs to look at Present and
Functional to determine state and health. This is how
other resources do it.
It should not look at totalCores to determine present / State.

HealthPopulate is not used. We use Functional to determine health status.
Functional is a bool that HealthPopulate cannot handle.

Signed-off-by: Chicago Duan <duanzhijia01@inspur.com>
Change-Id: Iff900061254fdd991b2dfa573b364a04decd8f98