Add BMC health metrics
The phosphor-health-monitor repository is publishing a few metrics
related the health of the BMC on DBus. This makes them very easy
to work with from BMCWeb, and allows a user to know the state of
the BMC on the web user interface.

This change adds a route to enable a user to read BMC health
metrics using the GET method.

Tested:
 - Tested using evb-ast2500 image on QEMU
   - when the DBus objects from health-monitor is not available,
     an error is returned
   - when health-monitor is running, the metrics it provides are
     correctly returned

Signed-off-by: Sui Chen <suichen6@gmail.com>
Change-Id: If8bd09140450a6104a5410b69ec33faeca26eca3