Delete escapePathForDbus
This method is incorrect per the dbus spec, and we should be using the
equivalent functions from sdbusplus, which were added as part of
fa0fa3b0fa5a9d59c37b6edbc53cc21b7c2a16b0
and
f87462af3571458fc8e33b087916bb1b77f38903

Note, the second patchset has not merged, and will be required for this
patchset to be merged.  Most changes are fairly mechanical in nature,
the only ones that aren't are the changes in manager.hpp, where dbus
escaping was being used for something that doesn't need dbus escaped,
which looks like a mistake in implementation.

Tested:
TBD, will try to test the parts that are available as I'm able.
Telemetry might be difficult, as it's new.

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: Ib4182b159c8361189a06b58c86712b1fd8f49f96