Update patch set 2
Patch Set 2:

The code you're editing is currently disabled:
https://github.com/openbmc/bmcweb/issues/188

If you have an interest in this dead code, please work on making efforts to get the backend upstreamed, and weigh in on the various issues here.  This code is currently slated for deletion if no backend is upstreamed.

Patch-set: 2
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>