Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> My 2 cents. First, there are possible (and real) implementations as design for such service is available in OpenBmc documentation.

We have different opinions on what "real" means.  A document is not real, it's one way an interface could hypothetically be built.  Only code that implements it is real, and today there is no code in OpenBMC that implements this.

>  Socket location is defined in DBus interface (Socket propoerty under (xyz.openbmc_project.VirtualMedia.MountPoint interace), and socket itself is created by bmcweb. My understanding is that ownership of socket directory is not bound to any component, but all parties shall assure that socket could be created.
> There is another issue, not mentioned before.  For security reasons socket shall be accessible only for authorized components.  Despite OpenBMC security model does not define access groups and almost all services are running as root, only root shall have access to socket (until special group and separate users are created).  And (what is important and not commonly known) socket permissions are defined by the permisions of the directory where it is created.  In summary, there shall be specific directory with approriate access rights created exclusively for sockets (using umask or changing access rights after creating).

I see a lot of things here that are true, but they're not really leading to anything actionable in the context of this patchset.  If you're interested in privilege isolation, you should be reviewing and testing the patches @mouse is pushing currently.  In the context of this patch, which is fixing a bug, I'm not sure what changes you're asking for.

I'm still of the opinion that we need to sort out the development of the virtual media backend, and it's merge with the existing one before any of these patches go through.

Patch-set: 2
Reviewer: Gerrit User 1000153 <1000153@adbd0a64-1f21-4d83-b585-671fe73cb6e4>